+++ bug_localization/src/utils/file_utils.py
