+++ bug_localization/src/utils/git_utils.py
